Create LSTM.py
This example formats data for an LSTM network by normalizing the input data and partitioning the data to a training and testing set. From here, a LSTM structure is created and trained using the partitioned training set data. Finally, the testing set data is used to create results for the network which are plotted and compared with actual results for the timeframe for accuracy using MAPE (Mean Absolute Percentage Error).